refactor: move typescript code next to rust code
Neither works without the presence of the other and this way they are
more easily found
