+++ packages/cargo_tools_vscode/src/extension/tasks/pinned/command.rs
@@ -1,5 +1,5 @@
 use futures::channel::mpsc::Sender;
-use wasm_bindgen::prelude::wasm_bindgen;
+use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::js_sys::Array;
 
 use crate::{
@@ -10,7 +10,9 @@ use crate::{
     },
 };
 
-#[wasm_bindgen(raw_module = "../cargoMakeTreeProvider.ts")]
+#[wasm_bindgen(
+    raw_module = "../../../packages/cargo_tools_vscode/src/extension/tasks/pinned/command.ts"
+)]
 extern "C" {
     #[wasm_bindgen]
     fn try_get_pinned_alias_key(value: Array) -> Option<String>;
