+++ packages/cargo_tools_vscode/src/extension/tasks/pinned/tree_provider.ts
@@ -1,0 +1,38 @@
+import * as vscode from 'vscode';
+import { CargoMakeNodeHandler, Icon } from '../../../../../../vscode_extension/src/wasm/cargo_tools_vscode';
+
+export class CargoMakePinnedNode extends vscode.TreeItem {
+    constructor(
+        public readonly label: string,
+        public readonly icon: Icon,
+        public readonly collapsibleState: vscode.TreeItemCollapsibleState,
+        public readonly contextValue: string,
+        public readonly description: string,
+        public readonly tooltip: string,
+        public readonly handler: CargoMakeNodeHandler,
+    ) {
+        super(label, collapsibleState);
+        this.iconPath = new vscode.ThemeIcon(icon.icon, new vscode.ThemeColor(icon.color));
+        this.contextValue = contextValue;
+        this.description = description;
+        this.tooltip = tooltip;
+        this.handler = handler;
+    }
+}
+
+export class PinnedAliasNode extends vscode.TreeItem {
+    constructor(
+        public readonly label: string,
+        public readonly icon: Icon,
+        public readonly collapsibleState: vscode.TreeItemCollapsibleState,
+        public readonly contextValue: string,
+        public readonly description: string,
+        public readonly tooltip: string,
+    ) {
+        super(label, collapsibleState);
+        this.iconPath = new vscode.ThemeIcon(icon.icon, new vscode.ThemeColor(icon.color));
+        this.contextValue = contextValue;
+        this.description = description;
+        this.tooltip = tooltip;
+    }
+}
