+++ packages/cargo_tools_vscode/src/extension/vscode_task_utils.ts
@@ -1,0 +1,6 @@
+import * as vscode from 'vscode';
+import { extension_context } from '../../../../vscode_extension/src/extension';
+
+export function register_command(command: string, callback: (args: any[]) => any) {
+    extension_context?.subscriptions.push(vscode.commands.registerCommand(command, (...args: any[]) => { return callback([...args]); }));
+}
