+++ vscode_extension/tsconfig.json
@@ -6,7 +6,7 @@
 			"ES2022"
 		],
 		"sourceMap": true,
-		"rootDir": "src",
+		"rootDir": "../",
 		"outDir": "out",
 		"strict": true,
 		"moduleResolution": "nodenext",
